pyproject.toml
[project]
name = "imagen-app"
version = "0.1.0"
description = "Image Generation App"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "huggingface-hub>=1.1.1",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
]
